--- Assets/Project/Scripts/MenuSelectScene/Settings/SettingsManager.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2816f2358c1b428e8a1ffb2a4bdc73c6
-timeCreated: 1587215494
