+++ Assets/Project/Scripts/Utils/PlayerPrefsUtils/UserSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db0def0d7a5374c8a9606472f0763769
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
